feat: Add session callback to options in options.ts
The code changes in this commit include adding a session callback to the options object in the options.ts file. This callback is responsible for modifying the session object based on the token received. In this case, the callback sets the user id in the session object if the token contains a 'sub' property. This change enhances the session handling functionality in the application.